requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
Pillow==9.0.0
pytube==11.0.2
requests==2.27.1
urllib3==1.26.8
